add routers to app add check cardId for integer add validation, getMediaByPostId to media controller change getPostBySlug to general not found or not published add getMediaById to database improte escapeInputMiddleware logic add uploadMiddleware.js fix authRoutes, postRoutes and adminroutes to use app.use(/route) add UploadRoutes - not implemented rework tests add postController.test.js change escaoeAllStrings to use DI domPurifyInstance add BildHochladen.js - as a memory file

+++ controllers/cardController.js
@@ -14,7 +14,6 @@ const createCard = async (cardData) => {
         throw error;
     }
 }
-
 const getAllCards = async () => {
     try {
         const cards = await DatabaseService.getAllCards();
@@ -37,9 +36,11 @@ const getAllCards = async () => {
         throw error;
     }
 }
-
 const getCardById = async (id) => {
     try {
+        if (!Number.isInteger(id) || id <= 0) {
+            throw new Error('Invalid card ID');
+        }
         const card = await DatabaseService.getCardById(id);
         if (!card) {
             throw new Error('Card not found');
@@ -54,9 +55,11 @@ const getCardById = async (id) => {
         throw error;
     }
 }
-
 const deleteCard = async (id) => {
     try {
+        if (!Number.isInteger(id) || id <= 0) {
+            throw new Error('Invalid card ID');
+        }
         const deleted = await DatabaseService.deleteCard(id);
         if (!deleted) {
             throw new Error('Card not found or not deleted');
@@ -67,7 +70,6 @@ const deleteCard = async (id) => {
         throw error;
     }
 }
-
 export default {
     createCard,
     getAllCards,
